compute hosts deployment enhancment: set VFs mac addresses
an optional task of setting mac addresses for part or all the VFs
for the compute host deployment.
this will make the usage of the domain isolation API easier when the admin
needs to specify the mac address when calling interface create endpoint,
he can use the predefined mac address.